refactor: 투입금액 Validation 테스트 LottoMachineTest로 이동
