+++ src/test/java/utils/InputValidationTest.java
@@ -12,48 +12,9 @@ public class InputValidationTest {
 
     static InputValidation inputValidation = new InputValidation();
 
-    @Test
-    void 문자_1000을_입력받으면_숫자1000을_리턴한다() {
-        //given
-        String givenMoney = "1000";
-        //when
-        int result = inputValidation.checkGivenMoney(givenMoney);
-        //then
-        assertThat(result).isEqualTo(1000);
-    }
 
-    @Test
-    void 문자_100을_입력받으면_런타임에러를_리턴한다() {
-        //given
-        String givenMoney = "100";
-        //when
-        //then
-        assertThatThrownBy(() ->
-                inputValidation.checkGivenMoney(givenMoney)).
-                isInstanceOf(RuntimeException.class);
-    }
 
-    @Test
-    void 문자_가를_입력받으면_런타임에러를_리턴한다() {
-        //given
-        String givenMoney = "가";
-        //when
-        //then
-        assertThatThrownBy(() ->
-                inputValidation.checkGivenMoney(givenMoney)).
-                isInstanceOf(RuntimeException.class);
-    }
 
-    @Test
-    void 문자_음수1000_입력받으면_런타임에러를_리턴한다() {
-        //given
-        String givenMoney = "-1000";
-        //when
-        //then
-        assertThatThrownBy(() ->
-                inputValidation.checkGivenMoney(givenMoney)).
-                isInstanceOf(RuntimeException.class);
-    }
 
     @Test
     void _1에서_45_사이의_로또보너스번호_입력_하면_숫자로_리턴() {
